feat: implement cart route, refactor header icons
