+++ src/views/basket/Basket.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function Basket() {
+  return (
+    <div>Basket</div>
+  )
+}
+
+export default Basket
